fix(input-set): fix youtube input call too many time to get segments
