+++ src/types/youtube/youtube.type.ts
@@ -1,23 +1,42 @@
-export interface IYoutubeMetadata {
-    title: string;
-    duration: number;
-    author: string;
-    views: number;
-    thumbnailUrl: string;
-    embed:
-        | {
-              iframe_url: string;
-              flash_url: string;
-              flash_secure_url: string;
-              width: any;
-              height: any;
-          }
-        | null
-        | undefined;
+interface IYoutubeCaptionSegment {
+    text: string;
+    startSecond: number;
+    startMs: number;
+    endSecond?: number;
+    endMs?: number;
+    duration?: number;
 }
 
-export interface IBalancedSegment {
+interface IEmbedVideoInfo {
+    iframe_url: string;
+    flash_url: string;
+    flash_secure_url: string;
+    width: number;
+    height: number;
+}
+
+interface VideoInfo {
+    title: string;
+    thumbnailUrl: string;
+    videoId: string;
+    duration: number;
+    embed: IEmbedVideoInfo;
+}
+
+interface YoutubeResourceMetadata {
+    url: string;
+    videoId: string;
+    videoInfo: VideoInfo;
+    content: string | IBalancedSegment[];
+    lengthContent: number;
+    wordCount: number;
+    segments: IYoutubeCaptionSegment[];
+}
+
+interface IBalancedSegment {
     startTime: number;
     endTime: number;
     text: string;
 }
+
+export { YoutubeResourceMetadata, IYoutubeCaptionSegment, IBalancedSegment, VideoInfo };
